build: split out the comments subpackage to subdir
