+++ packages/core/src/comments/index.ts
@@ -1,1 +1,9 @@
-export * from "../extensions/Comments/index.js";
+export * from "./models/User.js";
+export * from "./threadstore/DefaultThreadStoreAuth.js";
+export * from "./threadstore/ThreadStore.js";
+export * from "./threadstore/ThreadStoreAuth.js";
+export * from "./threadstore/TipTapThreadStore.js";
+export * from "./threadstore/yjs/RESTYjsThreadStore.js";
+export * from "./threadstore/yjs/YjsThreadStore.js";
+export * from "./threadstore/yjs/YjsThreadStoreBase.js";
+export * from "./types.js";
