+++ packages/core/src/comments/models/User.ts
@@ -1,3 +1,6 @@
+/**
+ * A collaborator of the document.
+ */
 export type User = {
   id: string;
   username: string;
