--- packages/core/src/extensions/Comments/index.ts
@@ -1,9 +1,0 @@
-export * from "./types.js";
-export * from "./threadstore/DefaultThreadStoreAuth.js";
-export * from "./threadstore/ThreadStore.js";
-export * from "./threadstore/ThreadStoreAuth.js";
-export * from "./threadstore/TipTapThreadStore.js";
-export * from "./threadstore/yjs/RESTYjsThreadStore.js";
-export * from "./threadstore/yjs/YjsThreadStore.js";
-export * from "./threadstore/yjs/YjsThreadStoreBase.js";
-export * from "../../models/User.js";
