+++ packages/core/src/extensions/Comments/userstore/UserStore.ts
@@ -1,4 +1,4 @@
-import { User } from "../../../models/User.js";
+import type { User } from "../../../comments/index.js";
 import { EventEmitter } from "../../../util/EventEmitter.js";
 
 /**
